+++ src/ops/cloud/AdminFederationOps.ts
@@ -11,6 +11,121 @@ import {
 import { getMetadata } from '../utils/ExportImportUtils';
 import { debugMessage } from '../utils/Console';
 import { getConfigEntity, putConfigEntity } from '../../api/IdmConfigApi';
+import State from '../../shared/State';
+
+export default class AdminFederationOps {
+  state: State;
+  constructor(state: State) {
+    this.state = state;
+  }
+
+  /**
+   * Create an empty idp export template
+   * @returns {AdminFederationExportInterface} an empty idp export template
+   */
+  createAdminFederationExportTemplate(): AdminFederationExportInterface {
+    return createAdminFederationExportTemplate({ state: this.state });
+  }
+
+  /**
+   * Get all admin federation providers
+   * @returns {Promise} a promise that resolves to an object containing an array of admin federation providers
+   */
+  async getAdminFederationProviders() {
+    return getAdminFederationProviders({ state: this.state });
+  }
+
+  /**
+   * Get admin federation provider by id
+   * @param {String} providerId social identity provider id/name
+   * @returns {Promise} a promise that resolves a social admin federation object
+   */
+  async getAdminFederationProvider(providerId: string) {
+    return getAdminFederationProvider({ providerId, state: this.state });
+  }
+
+  async putProviderByTypeAndId(
+    providerType: string,
+    providerId: string,
+    providerData: SocialIdpSkeleton
+  ) {
+    return putProviderByTypeAndId({
+      providerType,
+      providerId,
+      providerData,
+      state: this.state,
+    });
+  }
+
+  /**
+   * Delete admin federation provider by id
+   * @param {String} providerId admin federation provider id/name
+   * @returns {Promise} a promise that resolves to an admin federation provider object
+   */
+  async deleteAdminFederationProvider(
+    providerId: string
+  ): Promise<SocialIdpSkeleton> {
+    return deleteAdminFederationProvider({ providerId, state: this.state });
+  }
+
+  /**
+   * Export admin federation provider by id
+   * @param {string} providerId provider id/name
+   * @returns {Promise<AdminFederationExportInterface>} a promise that resolves to a SocialProviderExportInterface object
+   */
+  async exportAdminFederationProvider(
+    providerId: string
+  ): Promise<AdminFederationExportInterface> {
+    return exportAdminFederationProvider({ providerId, state: this.state });
+  }
+
+  /**
+   * Export all providers
+   * @returns {Promise<AdminFederationExportInterface>} a promise that resolves to a SocialProviderExportInterface object
+   */
+  async exportAdminFederationProviders(): Promise<AdminFederationExportInterface> {
+    return exportAdminFederationProviders({ state: this.state });
+  }
+
+  /**
+   * Import admin federation provider by id/name
+   * @param {string} providerId provider id/name
+   * @param {AdminFederationExportInterface} importData import data
+   */
+  async importAdminFederationProvider(
+    providerId: string,
+    importData: AdminFederationExportInterface
+  ): Promise<SocialIdpSkeleton> {
+    return importAdminFederationProvider({
+      providerId,
+      importData,
+      state: this.state,
+    });
+  }
+
+  /**
+   * Import first provider
+   * @param {AdminFederationExportInterface} importData import data
+   */
+  async importFirstAdminFederationProvider(
+    importData: AdminFederationExportInterface
+  ): Promise<SocialIdpSkeleton> {
+    return importFirstAdminFederationProvider({
+      importData,
+      state: this.state,
+    });
+  }
+
+  /**
+   * Import all providers
+   * @param {AdminFederationExportInterface} importData import data
+   */
+  async importAdminFederationProviders(
+    importData: AdminFederationExportInterface
+  ): Promise<SocialIdpSkeleton[]> {
+    return importAdminFederationProviders({ importData, state: this.state });
+  }
+}
 
 export interface AdminFederationExportInterface {
   meta?: ExportMetaData;
@@ -24,9 +139,13 @@ const ADMIN_FED_CONFIG_ID_PREFIX = 'fidc/federation-';
  * Create an empty idp export template
  * @returns {AdminFederationExportInterface} an empty idp export template
  */
-function createAdminFederationExportTemplate(): AdminFederationExportInterface {
+export function createAdminFederationExportTemplate({
+  state,
+}: {
+  state: State;
+}): AdminFederationExportInterface {
   return {
-    meta: getMetadata(),
+    meta: getMetadata({ state }),
     config: {},
     idp: {},
   } as AdminFederationExportInterface;
@@ -36,8 +155,8 @@ function createAdminFederationExportTemplate(): AdminFederationExportInterface {
  * Get all admin federation providers
  * @returns {Promise} a promise that resolves to an object containing an array of admin federation providers
  */
-export async function getAdminFederationProviders() {
-  const { result } = await _getAdminFederationProviders();
+export async function getAdminFederationProviders({ state }: { state: State }) {
+  const { result } = await _getAdminFederationProviders({ state });
   return result;
 }
 
@@ -46,8 +165,14 @@ export async function getAdminFederationProviders() {
  * @param {String} providerId social identity provider id/name
  * @returns {Promise} a promise that resolves a social admin federation object
  */
-export async function getAdminFederationProvider(providerId) {
-  const response = await getAdminFederationProviders();
+export async function getAdminFederationProvider({
+  providerId,
+  state,
+}: {
+  providerId: string;
+  state: State;
+}) {
+  const response = await getAdminFederationProviders({ state });
   const foundProviders = response.filter(
     (provider) => provider._id === providerId
   );
@@ -63,19 +188,32 @@ export async function getAdminFederationProvider(providerId) {
   }
 }
 
-export async function putProviderByTypeAndId(
-  providerType: string,
-  providerId: string,
-  providerData: object
-) {
-  debugMessage(`AdminFederationOps.putProviderByTypeAndId: start`);
+export async function putProviderByTypeAndId({
+  providerType,
+  providerId,
+  providerData,
+  state,
+}: {
+  providerType: string;
+  providerId: string;
+  providerData: SocialIdpSkeleton;
+  state: State;
+}) {
+  debugMessage({
+    message: `AdminFederationOps.putProviderByTypeAndId: start`,
+    state,
+  });
   try {
-    const response = await _putProviderByTypeAndId(
+    const response = await _putProviderByTypeAndId({
       providerType,
       providerId,
-      providerData
-    );
-    debugMessage(`AdminFederationOps.putProviderByTypeAndId: end`);
+      providerData,
+      state,
+    });
+    debugMessage({
+      message: `AdminFederationOps.putProviderByTypeAndId: end`,
+      state,
+    });
     return response;
   } catch (importError) {
     if (
@@ -86,18 +224,23 @@ export async function putProviderByTypeAndId(
       validAttributes.push('_id', '_type');
       for (const attribute of Object.keys(providerData)) {
         if (!validAttributes.includes(attribute)) {
-          debugMessage(`Removing invalid attribute: ${attribute}`);
+          debugMessage({
+            message: `Removing invalid attribute: ${attribute}`,
+            state,
+          });
           delete providerData[attribute];
         }
       }
-      const response = await _putProviderByTypeAndId(
+      const response = await _putProviderByTypeAndId({
         providerType,
         providerId,
-        providerData
-      );
-      debugMessage(
-        `AdminFederationOps.putProviderByTypeAndId: end (after retry)`
-      );
+        providerData,
+        state,
+      });
+      debugMessage({
+        message: `AdminFederationOps.putProviderByTypeAndId: end (after retry)`,
+        state,
+      });
       return response;
     } else {
       // re-throw unhandleable error
@@ -109,21 +252,26 @@ export async function putProviderByTypeAndId(
 /**
  * Delete admin federation provider by id
  * @param {String} providerId admin federation provider id/name
- * @returns {Promise} a promise that resolves to an admin federation provider object
+ * @returns {Promise<SocialIdpSkeleton>} a promise that resolves to an admin federation provider object
  */
-export async function deleteAdminFederationProvider(
-  providerId: string
-): Promise<unknown> {
-  const response = await getAdminFederationProviders();
+export async function deleteAdminFederationProvider({
+  providerId,
+  state,
+}: {
+  providerId: string;
+  state: State;
+}): Promise<SocialIdpSkeleton> {
+  const response = await getAdminFederationProviders({ state });
   const foundProviders = response.filter(
     (provider) => provider._id === providerId
   );
   switch (foundProviders.length) {
     case 1:
-      return await deleteProviderByTypeAndId(
-        foundProviders[0]._type._id,
-        foundProviders[0]._id
-      );
+      return await deleteProviderByTypeAndId({
+        providerType: foundProviders[0]._type._id,
+        providerId: foundProviders[0]._id,
+        state,
+      });
     case 0:
       throw new Error(`Provider '${providerId}' not found`);
     default:
@@ -138,18 +286,26 @@ export async function deleteAdminFederationProvider(
  * @param {string} providerId provider id/name
  * @returns {Promise<AdminFederationExportInterface>} a promise that resolves to a SocialProviderExportInterface object
  */
-export async function exportAdminFederationProvider(
-  providerId: string
-): Promise<AdminFederationExportInterface> {
-  debugMessage(`AdminFederationOps.exportAdminFederationProvider: start`);
-  const exportData = createAdminFederationExportTemplate();
+export async function exportAdminFederationProvider({
+  providerId,
+  state,
+}: {
+  providerId: string;
+  state: State;
+}): Promise<AdminFederationExportInterface> {
+  debugMessage({
+    message: `AdminFederationOps.exportAdminFederationProvider: start`,
+    state,
+  });
+  const exportData = createAdminFederationExportTemplate({ state });
   const errors = [];
   try {
-    const idpData = await getAdminFederationProvider(providerId);
+    const idpData = await getAdminFederationProvider({ providerId, state });
     exportData.idp[idpData._id] = idpData;
-    const idpConfig = await getConfigEntity(
-      `${ADMIN_FED_CONFIG_ID_PREFIX}${providerId}`
-    );
+    const idpConfig = await getConfigEntity({
+      entityId: `${ADMIN_FED_CONFIG_ID_PREFIX}${providerId}`,
+      state,
+    });
     exportData.config[idpConfig._id] = idpConfig;
   } catch (error) {
     errors.push(error);
@@ -158,7 +314,10 @@ export async function exportAdminFederationProvider(
     const errorMessages = errors.map((error) => error.message).join('\n');
     throw new Error(`Export error:\n${errorMessages}`);
   }
-  debugMessage(`AdminFederationOps.exportAdminFederationProvider: end`);
+  debugMessage({
+    message: `AdminFederationOps.exportAdminFederationProvider: end`,
+    state,
+  });
   return exportData;
 }
 
@@ -166,18 +325,26 @@ export async function exportAdminFederationProvider(
  * Export all providers
  * @returns {Promise<AdminFederationExportInterface>} a promise that resolves to a SocialProviderExportInterface object
  */
-export async function exportAdminFederationProviders(): Promise<AdminFederationExportInterface> {
-  debugMessage(`AdminFederationOps.exportAdminFederationProviders: start`);
-  const exportData = createAdminFederationExportTemplate();
+export async function exportAdminFederationProviders({
+  state,
+}: {
+  state: State;
+}): Promise<AdminFederationExportInterface> {
+  debugMessage({
+    message: `AdminFederationOps.exportAdminFederationProviders: start`,
+    state,
+  });
+  const exportData = createAdminFederationExportTemplate({ state });
   const errors = [];
   try {
-    const allIdpsData = await getAdminFederationProviders();
+    const allIdpsData = await getAdminFederationProviders({ state });
     for (const idpData of allIdpsData) {
       try {
         exportData.idp[idpData._id] = idpData;
-        const idpConfig = await getConfigEntity(
-          `${ADMIN_FED_CONFIG_ID_PREFIX}${idpData._id}`
-        );
+        const idpConfig = await getConfigEntity({
+          entityId: `${ADMIN_FED_CONFIG_ID_PREFIX}${idpData._id}`,
+          state,
+        });
         exportData.config[idpConfig._id] = idpConfig;
       } catch (error) {
         errors.push(error);
@@ -190,7 +357,10 @@ export async function exportAdminFederationProviders(): Promise<AdminFederationE
     const errorMessages = errors.map((error) => error.message).join('\n');
     throw new Error(`Export error:\n${errorMessages}`);
   }
-  debugMessage(`AdminFederationOps.exportAdminFederationProviders: end`);
+  debugMessage({
+    message: `AdminFederationOps.exportAdminFederationProviders: end`,
+    state,
+  });
   return exportData;
 }
 
@@ -199,24 +369,34 @@ export async function exportAdminFederationProviders(): Promise<AdminFederationE
  * @param {string} providerId provider id/name
  * @param {AdminFederationExportInterface} importData import data
  */
-export async function importAdminFederationProvider(
-  providerId: string,
-  importData: AdminFederationExportInterface
-): Promise<SocialIdpSkeleton> {
+export async function importAdminFederationProvider({
+  providerId,
+  importData,
+  state,
+}: {
+  providerId: string;
+  importData: AdminFederationExportInterface;
+  state: State;
+}): Promise<SocialIdpSkeleton> {
   let response = null;
   const errors = [];
   const imported = [];
   for (const idpId of Object.keys(importData.idp)) {
     if (idpId === providerId) {
       try {
-        response = await putProviderByTypeAndId(
-          importData.idp[idpId]._type._id,
-          idpId,
-          importData.idp[idpId]
-        );
+        response = await putProviderByTypeAndId({
+          providerType: importData.idp[idpId]._type._id,
+          providerId: idpId,
+          providerData: importData.idp[idpId],
+          state,
+        });
         const configId = `${ADMIN_FED_CONFIG_ID_PREFIX}${idpId}`;
         if (importData.config[configId]) {
-          await putConfigEntity(configId, importData.config[configId]);
+          await putConfigEntity({
+            entityId: configId,
+            entityData: importData.config[configId],
+            state,
+          });
         }
         imported.push(idpId);
       } catch (error) {
@@ -238,22 +418,31 @@ export async function importAdminFederationProvider(
  * Import first provider
  * @param {AdminFederationExportInterface} importData import data
  */
-export async function importFirstAdminFederationProvider(
-  importData: AdminFederationExportInterface
-): Promise<SocialIdpSkeleton> {
+export async function importFirstAdminFederationProvider({
+  importData,
+  state,
+}: {
+  importData: AdminFederationExportInterface;
+  state: State;
+}): Promise<SocialIdpSkeleton> {
   let response = null;
   const errors = [];
   const imported = [];
   for (const idpId of Object.keys(importData.idp)) {
     try {
-      response = await putProviderByTypeAndId(
-        importData.idp[idpId]._type._id,
-        idpId,
-        importData.idp[idpId]
-      );
+      response = await putProviderByTypeAndId({
+        providerType: importData.idp[idpId]._type._id,
+        providerId: idpId,
+        providerData: importData.idp[idpId],
+        state,
+      });
       const configId = `${ADMIN_FED_CONFIG_ID_PREFIX}${idpId}`;
       if (importData.config[configId]) {
-        await putConfigEntity(configId, importData.config[configId]);
+        await putConfigEntity({
+          entityId: configId,
+          entityData: importData.config[configId],
+          state,
+        });
       }
       imported.push(idpId);
     } catch (error) {
@@ -274,24 +463,33 @@ export async function importFirstAdminFederationProvider(
  * Import all providers
  * @param {AdminFederationExportInterface} importData import data
  */
-export async function importAdminFederationProviders(
-  importData: AdminFederationExportInterface
-): Promise<SocialIdpSkeleton[]> {
+export async function importAdminFederationProviders({
+  importData,
+  state,
+}: {
+  importData: AdminFederationExportInterface;
+  state: State;
+}): Promise<SocialIdpSkeleton[]> {
   const response = [];
   const errors = [];
   const imported = [];
   for (const idpId of Object.keys(importData.idp)) {
     try {
       response.push(
-        await putProviderByTypeAndId(
-          importData.idp[idpId]._type._id,
-          idpId,
-          importData.idp[idpId]
-        )
+        await putProviderByTypeAndId({
+          providerType: importData.idp[idpId]._type._id,
+          providerId: idpId,
+          providerData: importData.idp[idpId],
+          state,
+        })
       );
       const configId = `${ADMIN_FED_CONFIG_ID_PREFIX}${idpId}`;
       if (importData.config[configId]) {
-        await putConfigEntity(configId, importData.config[configId]);
+        await putConfigEntity({
+          entityId: configId,
+          entityData: importData.config[configId],
+          state,
+        });
       }
       imported.push(idpId);
     } catch (error) {